Change the values or settings of a dropzone on the client

The set of presets can be cleared by using presets=character(0).

updateDropZoneInput(
  session = getDefaultReactiveDomain(),
  inputId,
  presets = NULL,
  choices = NULL,
  placeholder = NULL
)

Arguments

session
The session object passed to function given to shinyServer.

inputId
The input slot that will be used to acces the value.

presets
Array or list of preset values.

choices
List of acceptable values with their associated labels. Note that the labels can be arbitrary HTML, as long as they are wrapped in a tagList.

placeholder
Insert placeholder text.